requirements.txt
streamlit
face_recognition
Pillow
uuid
requests
gdown
dotenv